requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cssselect==1.2.0
cssutils==2.11.0
DateTime==5.5
email-to==0.1.0
idna==3.7
install==1.3.5
lxml==5.2.2
Markdown==3.6
numpy==1.26.4
pandas==2.2.2
pathlib==1.0.1
premailer==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tzdata==2024.1
urllib3==2.0.2
zope.interface==6.4.post2
